+++ autoarray/plot/wrap/two_d/border_scatter.py
@@ -1,0 +1,11 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+
+class BorderScatter(GridScatter):
+    """
+    Plots a border over an image, using the `Mask2d` object's (y,x) `border_grid_sub_1` property.
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
