+++ autoarray/plot/wrap/two_d/index_scatter.py
@@ -1,0 +1,11 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+
+class IndexScatter(GridScatter):
+    """
+    Plots specific (y,x) coordinates of a grid (or grids) via their 1d or 2d indexes.
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
