+++ autoarray/plot/wrap/two_d/mask_scatter.py
@@ -1,0 +1,10 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+class MaskScatter(GridScatter):
+    """
+    Plots a mask over an image, using the `Mask2d` object's (y,x) `edge_grid_sub_1` property.
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
