+++ autoarray/plot/wrap/two_d/mesh_grid_scatter.py
@@ -1,0 +1,11 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+
+class MeshGridScatter(GridScatter):
+    """
+    Plots the grid of a `Mesh` object (see `autoarray.inversion`).
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
