+++ autoarray/plot/wrap/two_d/origin_scatter.py
@@ -1,0 +1,10 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+class OriginScatter(GridScatter):
+    """
+    Plots the (y,x) coordinates of the origin of a data structure (e.g. as a black cross).
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
