+++ autoarray/plot/wrap/two_d/parallel_overscan_plot.py
@@ -1,0 +1,5 @@
+from autoarray.plot.wrap.two_d.grid_plot import GridPlot
+
+
+class ParallelOverscanPlot(GridPlot):
+    pass
