+++ autoarray/plot/wrap/two_d/positions_scatter.py
@@ -1,0 +1,11 @@
+from autoarray.plot.wrap.two_d.grid_scatter import GridScatter
+
+
+class PositionsScatter(GridScatter):
+    """
+    Plots the (y,x) coordinates that are input in a plotter via the `positions` input.
+
+    See `mat_structure.Scatter` for a description of how matplotlib is wrapped to make this plot.
+    """
+
+    pass
