+++ autoarray/plot/wrap/two_d/serial_overscan_plot.py
@@ -1,0 +1,4 @@
+from autoarray.plot.wrap.two_d.grid_plot import GridPlot
+
+class SerialOverscanPlot(GridPlot):
+    pass
