+++ autoarray/plot/wrap/two_d/serial_prescan_plot.py
@@ -1,0 +1,5 @@
+from autoarray.plot.wrap.two_d.grid_plot import GridPlot
+
+
+class SerialPrescanPlot(GridPlot):
+    pass
